Mather_bots/requirements.txt
python-telegram-bot==20.7
psutil==5.9.6
schedule==1.2.0
python-dotenv==1.0.0
nest-asyncio==1.6.0 